requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
zipp==3.8.0
